pyproject.toml
[project]
name = "chat-app-nugen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.133.0",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
